// src/types/cart.ts
import { Product } from "./product";

export interface CartItem {
    product: Product;
    quantity: number;
}

export interface CartStore {
    items: CartItem[];
    addItem: (product: Product) => void;
    removeItem: (productId: string) => void;
    updateQuantity: (productId: string, quantity: number) => void;
    clearCart: () => void;
    total: number;
} 